Add files via upload
The analysis in this project is based on data extracted from the Gold Layer of the Data Warehouse I am currently developing as part of an ongoing academic and applied data project. The Gold Layer contains cleaned, enriched, and business-ready datasets, making it ideal for performing advanced exploratory data analysis (EDA) and deriving actionable insights.